docs: switch to BYOCNI in AKS prerequisites / instructions
Microsoft has recently released support for "Bring your own CNI" when
creating AKS clusters, allowing users to create clusters with no CNI
plugin pre-installed.

This is extremely useful for Cilium as previously setting up an AKS
cluster and installing Cilium was very complicated, due to the need to
use multiple nodepools with a complex taint system in order to ensure
applications pods would not get scheduled before Cilium agents were
ready to manage connectivity in the cluster.

The BYOCNI feature has been available in preview as an `az` CLI
extension for a couple of weeks now, and Microsoft has just published a
dedicated documentation page outlining the prequisites and implications
(notably in terms of support policy) of BYOCNI:
https://docs.microsoft.com/en-us/azure/aks/use-byo-cni?tabs=azure-cli

BYOCNI in its current state is a strong enough proposal to warrant a
change to the recommendations we make to our users for newly created AKS
clusters.

There are changes in terms of installing Cilium itself, though:

- The older Azure-specific operator image, previously required for Azure
  IPAM, is not working in BYOCNI mode.
- Instead, the generic operator image must be used and configured for
  cluster-pool IPAM.
- Using cluster-pool lifts the necessity of setting up an Azure Service
  Principal with necessary privileges over the node resource group
  specifically for IPAM.

Signed-off-by: Nicolas Busseneau <nicolas@isovalent.com>